BBDMS
Acne is a prevalent skin condition affecting millions worldwide, with significant psychosocial and dermatological impacts. Timely and accurate diagnosis of acne lesions is crucial for effective treatment and management. In this research paper, we propose a novel approach for automated acne detection leveraging Convolutional Neural Networks (CNNs), a powerful deep learning technique.
Furthermore, we compare the performance of our CNN-based approach with traditional image processing techniques and manual diagnosis by dermatologists. The results demonstrate the superior accuracy and efficiency of the CNN model in identifying acne lesions, showcasing its potential as a valuable tool for dermatological diagnosis and healthcare delivery.